Move non-parallel commands to separate files
