+++ clowder/clowder/cli/sync_controller.py
@@ -10,12 +10,15 @@ from __future__ import print_function
 from cement.ext.ext_argparse import ArgparseController, expose
 from termcolor import cprint
 
-from clowder.cli.parallel_commands import sync
+from clowder.cli.parallel_commands import sync_parallel
 from clowder.clowder_controller import CLOWDER_CONTROLLER
 from clowder.clowder_repo import print_clowder_repo_status_fetch
 from clowder.util.connectivity import network_connection_required
 from clowder.util.decorators import valid_clowder_yaml_required
-from clowder.util.clowder_utils import options_help_message
+from clowder.util.clowder_utils import (
+    filter_projects,
+    options_help_message
+)
 
 
 class SyncController(ArgparseController):
@@ -56,3 +59,23 @@ class SyncController(ArgparseController):
             cprint(' - No forks to sync\n', 'red')
             return
         sync(CLOWDER_CONTROLLER, all_fork_projects, rebase=self.app.pargs.rebase, parallel=self.app.pargs.parallel)
+
+
+def sync(clowder, project_names, rebase=False, parallel=False):
+    """Sync projects
+
+    .. py:function:: sync(clowder, project_names, rebase=False, parallel=False)
+
+    :param ClowderController clowder: ClowderController instance
+    :param list[str] project_names: Project names to sync
+    :param Optional[bool] rebase: Whether to use rebase instead of pulling latest changes
+    :param Optional[bool] parallel: Whether command is being run in parallel, affects output
+    """
+
+    projects = filter_projects(clowder.groups, project_names=project_names)
+    if parallel:
+        sync_parallel(projects, rebase=rebase)
+        return
+
+    for project in projects:
+        project.sync(rebase=rebase)
